Squashed a bug & added error handling - dupe_fixer.py
## Bug fix
There was a small bug where if the launcher accidentally writes a new line in between a duplicate error line, the program would still try to parse the line.
This was due to the way the program checked for error lines; it would check if any one of the key words were in that line. If the line was split in half by mistake of the launcher, it would see one of the keywords and attempt to parse the line, but would error on tokenising the line by a keyword because it's not in the line.
This has now been patched. 

## Error system implementation
I have also added a catch-all error system that writes any unhandled errors into a .log file in the cwd, and then informs the user of the error and that they need to send the log file to either me (prov3ntus) or Shidouri to be looked in to.